create SearchTitleAndWidget
