+++ src/features/search/components/SearchHeader/SearchHeader.web.test.tsx
@@ -52,6 +52,7 @@ describe('SearchHeader component', () => {
           searchInputID={searchInputID}
           addSearchHistory={jest.fn()}
           searchInHistory={jest.fn()}
+          searchView={SearchView.Landing}
         />
       )
       await act(async () => {})
@@ -66,6 +67,7 @@ describe('SearchHeader component', () => {
         searchInputID={searchInputID}
         addSearchHistory={jest.fn()}
         searchInHistory={jest.fn()}
+        searchView={SearchView.Landing}
       />
     )
 
@@ -86,6 +88,7 @@ describe('SearchHeader component', () => {
         searchInputID={searchInputID}
         addSearchHistory={jest.fn()}
         searchInHistory={jest.fn()}
+        searchView={SearchView.Landing}
       />
     )
     await act(async () => {})
@@ -105,6 +108,7 @@ describe('SearchHeader component', () => {
           searchInputID={searchInputID}
           addSearchHistory={jest.fn()}
           searchInHistory={jest.fn()}
+          searchView={view}
         />
       )
 
@@ -125,6 +129,7 @@ describe('SearchHeader component', () => {
         searchInputID={searchInputID}
         addSearchHistory={jest.fn()}
         searchInHistory={jest.fn()}
+        searchView={SearchView.Landing}
       />
     )
